backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.40,<3
python-dotenv==1.0.0
pydantic>=2.5,<3
psycopg[binary]>=3.1.18
